Use Doubles in the american example.
My laptop doesn't have enough memory for the larger instances after the switch
from Float to Double, so I've also reduced the largest instance size.

+++ examples/american/American/Nikola.hs
@@ -16,7 +16,7 @@ import Data.Array.Nikola.Backend.CUDA
 -- American options available at <http://github.com/kfl/american-options>.
 --
 
-type F = Float
+type F = Double
 
 v1 ^*^ v2 = zipWith (*) v1 v2
 v1 ^+^ v2 = zipWith (+) v1 v2
